\name{semiBase-package}
\alias{semiBase-package}
\alias{semiBase}
\docType{package}
\title{
\packageTitle{semiBase}
}
\description{
\packageDescription{semiBase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{semiBase}
\packageIndices{semiBase}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{semiBase}

Maintainer: \packageMaintainer{semiBase}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
